pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-setup-forge"
version = "0.1.0"
description = "Manage, install, and remove agent skills and agent definitions for Claude Code, Mistral Vibe, and Copilot CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["ai", "agent", "skills", "claude-code", "copilot", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "python-frontmatter>=1.1",
    "httpx>=0.27",
]

[project.urls]
Homepage = "https://github.com/jyjeanne/ai-setup-forge"
Repository = "https://github.com/jyjeanne/ai-setup-forge"
Issues = "https://github.com/jyjeanne/ai-setup-forge/issues"

[project.scripts]
ai-setup-forge = "ai_setup_forge.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_setup_forge"]

[tool.hatch.build.targets.wheel.force-include]
"skills" = "ai_setup_forge/bundled_skills"
"agents" = "ai_setup_forge/bundled_agents"

[tool.hatch.build.targets.sdist]
include = ["src/ai_setup_forge", "skills", "agents", "tests", "docs", "README.md", "pyproject.toml"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
